backend/requirements.txt
fastapi
uvicorn
pydantic
requests
python-multipart
ultralytics
numpy
opencv-python-headless
pillow
python-dotenv